Adiciona confirmação de exclusão

+++ src/main/java/br/senac/pi/TelaConsultaClienteController.java
@@ -8,6 +8,7 @@ import java.util.ResourceBundle;
 import javafx.event.ActionEvent;
 import javafx.fxml.FXML;
 import javafx.fxml.Initializable;
+import javafx.scene.control.Alert;
 import javafx.scene.control.TableColumn;
 import javafx.scene.control.TableView;
 import javafx.scene.control.TextField;
@@ -145,24 +146,32 @@ public class TelaConsultaClienteController implements Initializable {
 
     @FXML
     private void excluirCliente(ActionEvent event) {
-        LinhaTabelaCliente linha = tabelaCliente.getSelectionModel().getSelectedItem();
         
-        if (linha != null) {
-            int id = linha.getId();
+        if (App.perguntar("Exclusão", "Deseja mesmo excluir?", "A escolha não pode ser desfeita!")) {
             
-            String sql = "delete from cliente where id_cliente = ?";
-            
-            try (PreparedStatement ps = DB.connect().prepareStatement(sql)){
-                ps.setInt(1, id);
-                
-                ps.execute();
-                
-                listar();
-                
-            } catch (Exception e) {
-                e.printStackTrace();
+            LinhaTabelaCliente linha = tabelaCliente.getSelectionModel().getSelectedItem();
+
+            if (linha != null) {
+                int id = linha.getId();
+
+                String sql = "delete from cliente where id_cliente = ?";
+
+                try (PreparedStatement ps = DB.connect().prepareStatement(sql)){
+                    ps.setInt(1, id);
+
+                    ps.execute();
+
+                    listar();
+
+                } catch (Exception e) {
+                    e.printStackTrace();
+                }
+            }else{
+                Alert alert = new Alert(Alert.AlertType.ERROR);
+                alert.setContentText("Nenhum cliente selecionado");
+                alert.showAndWait();
             }
-        } 
+        }
     }
 
     @FXML
